app/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
pydantic
python-jose[cryptography]
sqlalchemy
python-multipart
